using kafka client.py
