+++ src/piddiplatsch/cli.py
@@ -2,58 +2,15 @@ import click
 import os
 import logging
 import json
-import uuid
-from confluent_kafka import Producer
-from confluent_kafka.admin import AdminClient, NewTopic
 from piddiplatsch.consumer import Consumer, process_message
 from piddiplatsch.config import config
+from piddiplatsch import client
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
 
-def get_producer(kafka_server):
-    return Producer({"bootstrap.servers": kafka_server})
-
-
-def get_admin_client(kafka_server):
-    return AdminClient({"bootstrap.servers": kafka_server})
-
-
-def ensure_topic_exists(kafka_server, topic):
-    admin_client = get_admin_client(kafka_server)
-    metadata = admin_client.list_topics(timeout=5)
-    if topic not in metadata.topics:
-        click.echo(f"ℹ️ Topic '{topic}' does not exist. Creating it...")
-        new_topic = NewTopic(topic, num_partitions=1, replication_factor=1)
-        fs = admin_client.create_topics([new_topic])
-        try:
-            fs[topic].result()
-            click.echo(f"✅ Created Kafka topic: {topic}")
-        except Exception as e:
-            click.echo(f"⚠️  Failed to create topic '{topic}': {e}", err=True)
-
-
-def produce_message(kafka_server, topic, key, value):
-    ensure_topic_exists(kafka_server, topic)
-    producer = get_producer(kafka_server)
-
-    def delivery_report(err, msg):
-        if err:
-            click.echo(f"❌ Delivery failed: {err}", err=True)
-        else:
-            click.echo(f"📤 Message delivered to {msg.topic()} [{msg.partition()}]")
-
-    producer.produce(
-        topic,
-        key=key.encode("utf-8"),
-        value=value.encode("utf-8"),
-        callback=delivery_report,
-    )
-    producer.flush()
-
-
 def start_kafka_consumer(topic, kafka_server):
-    ensure_topic_exists(kafka_server, topic)
+    client.ensure_topic_exists(kafka_server, topic)
     logging.info(f"Starting Kafka consumer for topic: {topic}")
     consumer = Consumer(topic, kafka_server)
     for key, value in consumer.consume():
@@ -127,32 +84,30 @@ def send(ctx, message, path, topic, kafka_server, verbose):
         click.echo("❌ Please provide a message or a path to a JSON file.", err=True)
         ctx.exit(1)
 
-    if path:
-        try:
-            with open(path, "r") as f:
-                data = json.load(f)
-            key = os.path.splitext(os.path.basename(path))[0]
-            value = json.dumps(data)
-        except FileNotFoundError:
-            click.echo(f"❌ File not found: {path}", err=True)
-            ctx.exit(1)
-        except json.JSONDecodeError as e:
-            click.echo(f"❌ Invalid JSON in file: {e}", err=True)
-            ctx.exit(1)
-    else:
-        try:
-            data = json.loads(message)
-            value = json.dumps(data)
-        except json.JSONDecodeError as e:
-            click.echo(f"❌ Invalid JSON in --message: {e}", err=True)
-            ctx.exit(1)
-        key = str(uuid.uuid5(uuid.NAMESPACE_DNS, message))
+    try:
+        if path:
+            key, value = client.build_message_from_path(path)
+        else:
+            key, value = client.build_message_from_json_string(message)
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        click.echo(f"❌ {e}", err=True)
+        ctx.exit(1)
 
     if verbose:
         click.echo(f"🔑 Key: {key}")
         click.echo(f"📦 Value: {value}")
 
-    produce_message(kafka_server, topic, key, value)
+    def report(err, msg):
+        if err:
+            click.echo(f"❌ Delivery failed: {err}", err=True)
+        else:
+            click.echo(f"📤 Message delivered to {msg.topic()} [{msg.partition()}]")
+
+    try:
+        client.send_message(kafka_server, topic, key, value, on_delivery=report)
+    except Exception as e:
+        click.echo(f"❌ {e}", err=True)
+        ctx.exit(1)
 
 
 if __name__ == "__main__":
